# openapi/delta-backend.openapi.yaml
openapi: 3.0.0
paths:
...
  /api/business/catalog/documents:
    post:
      operationId: BusinessCatalog_createDocument
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - company_name
                - document_name
                - file
              properties:
                company_name:
                  type: string
                category_name:
                  type: string
                document_name:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - businessSession: []
          csrfToken: []
      summary: Upload a catalog PDF document
      tags:
        - Business Catalog
...
  /api/business/catalog/documents/{document_id}:
    put:
      operationId: BusinessCatalog_updateDocument
      parameters:
        - name: document_id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document_name:
                  type: string
                category_name:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - businessSession: []
          csrfToken: []
      summary: Update or replace a catalog PDF document
      tags:
        - Business Catalog
    delete:
      operationId: BusinessCatalog_deleteDocument
      parameters:
        - name: document_id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - businessSession: []
          csrfToken: []
      summary: Delete a catalog document
      tags:
        - Business Catalog
  /api/business/catalog/companies/{company_slug}:
    put:
      operationId: BusinessCatalog_updateCompany
      parameters:
        - name: company_slug
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogCompanyDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - businessSession: []
          csrfToken: []
      summary: Rename or merge a catalog company
      tags:
        - Business Catalog
...
  /api/categories:
    get:
      operationId: Product_getCategories
      parameters:
        - name: companyId
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: List active product categories
      tags:
        - Products
  /api/products:
    get:
      operationId: Product_getProducts
      parameters:
        - name: companyId
          required: false
          in: query
          schema:
            type: string
        - name: categoryId
          required: false
          in: query
          schema:
            type: string
        - name: subcategoryId
          required: false
          in: query
          schema:
            type: string
        - name: stock
          required: false
          in: query
          schema:
            type: string
            enum:
              - in_stock
              - out_of_stock
        - name: search
          required: false
          in: query
          schema:
            maxLength: 120
            type: string
        - name: sort
          required: true
          in: query
          schema:
            default: name_asc
            type: string
            enum:
              - name_asc
              - name_desc
              - price_asc
              - price_desc
        - name: limit
          required: true
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 30
            allOf:
              - $ref: "#/components/schemas/Object"
        - name: cursor
          required: false
          in: query
          schema:
            maxLength: 2048
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: List and filter active products
      tags:
        - Products
  /api/products/{productId}:
    get:
      operationId: Product_getProduct
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Get active product details
      tags:
        - Products
  /api/internal/catalog/documents:
    post:
      operationId: InternalCatalog_createDocument
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - company_name
                - document_name
                - file
              properties:
                company_name:
                  type: string
                category_name:
                  type: string
                document_name:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - internalAdmin: []
      summary: Upload a catalog PDF through the internal API
      tags:
        - Internal Catalog
  /api/internal/catalog/documents/{document_id}:
    put:
      operationId: InternalCatalog_updateDocument
      parameters:
        - name: document_id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document_name:
                  type: string
                category_name:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - internalAdmin: []
      summary: Update a catalog PDF through the internal API
      tags:
        - Internal Catalog
    delete:
      operationId: InternalCatalog_deleteDocument
      parameters:
        - name: document_id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - internalAdmin: []
      summary: Delete a catalog document through the internal API
      tags:
        - Internal Catalog
  /api/internal/catalog/companies/{company_slug}:
    put:
      operationId: InternalCatalog_updateCompany
      parameters:
        - name: company_slug
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogCompanyDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - internalAdmin: []
      summary: Rename a catalog company through the internal API
      tags:
        - Internal Catalog
...
info:
  title: Delta Backend API
  description: Public catalog and product APIs, catalog access flows,
    authenticated business administration, and internal administration
    endpoints.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    businessSession:
      type: apiKey
      in: cookie
      name: business_session
      description: Business session cookie. Production uses __Host-business_session.
    catalogAccess:
      type: apiKey
      in: cookie
      name: catalog_access
      description: Catalog access session cookie.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token returned by business authentication.
    internalAdmin:
      type: apiKey
      in: header
      name: X-Backend-Admin-Token
      description: Token for the optional internal administration API.
  schemas:
...
    UpdateCatalogCompanyDto:
      type: object
      properties:
        company_name:
          type: string
      required:
        - company_name
...
    Object:
      type: object
      properties: {}